Initial clean up of Spring Boot dependencies
The version was not up to date, plus there are many conflicts
between Spring Boot and FF4J parent, so we can't really use one
of those as a parent and the other as a BOM.

Also excludes a load of things that looked like they should be
optional in FF4J (like grizzly and jersey test jars).

And removed spring-boot-admin because it doesn't add anything to
the sample.